Fix issue regarding buffer not being properly set. Change hardware type to board purchased: will fix bitmaps resulting from this change
Fixed an issue regarding how the buffer was setting the "mx" object so it wouldn't create copies via Pass by Value. Additionally, changed the Hardware Type for the "mx" to the board being used, which resulted in flipped bitmap displays. Will fix these bitmaps in the next commit. Then, work on implementing the rest of the body part classes into the Header/Source file structure